move function to user.service
