+++ libs/user/src/lib/auth/+state/auth.service.ts
@@ -85,95 +85,4 @@ export class AuthService extends FireAuthService<AuthState> {
       surname: ctx.surname
     };
   }
-
-  //////////
-  // USER //
-  //////////
-
-  /** Call a firebase function to get or create a user.
-   * @email find the user with this email. If email doesn't match with an existing user,
-   * create a user with this email address.
-   */
-  public async getOrCreateUserByMail(email: string, orgName: string, invitationId?: string): Promise<User> {
-    const f = this.functions.httpsCallable('getOrCreateUserByMail');
-    return f({ email, orgName }).toPromise();
-  }
-
-  /**
-   * Check if uid is exists in blockframesAdmin collection.
-   * If document exists, user is blockframeAdmin (like an ancient god).
-   * @param uid 
-   */
-  public async isBlockframesAdmin(uid: string = this.query.userId): Promise<boolean> {
-    const snap = await this.db.collection('blockframesAdmin').doc(uid).get().toPromise();
-    return snap.exists;
-  }
-
-  /**
-   * Set/unset user as blockframesAdmin
-   * @param state
-   * @param uid 
-   */
-  public async setBlockframesAdmin(state: boolean = true, uid: string = this.query.userId): Promise<void> {
-    if (state) {
-      await this.db.collection('blockframesAdmin').doc(uid).set({});
-    } else {
-      await this.db.collection('blockframesAdmin').doc(uid).delete();
-    }
-  }
-
-  /**
-   * Checks if an user exists
-   * @dev If in the future, we need to keep an user list in the state other than members of an org, 
-   * this will be the time to create a userService and to move this method in it.
-   * @param uid
-   */
-  public async userExists(uid: string): Promise<boolean> {
-    const snap = await this.db.collection('users').doc(uid).get().toPromise();
-    return snap.exists;
-  }
-
-  /**
-   * Fetch an user based on his uid
-   * @dev If in the future, we need to keep an user list in the state other than members of an org, 
-   * this will be the time to create a userService and to move this method in it.
-   * @param uid
-   */
-  public async getUser(uid: string): Promise<User> {
-    const user = await this.db.collection('users').doc(uid).get().toPromise();
-    return user.data() as User;
-  }
-
-  /**
-   * @dev since this.update() does not behave like other services 
-   * (authService extends FireAuthService<AuthState> and others extends CollectionService<XxxState>)
-   * This method was created to easily update an user.
-   * @param uid 
-   * @param update 
-   */
-  public async updateById(uid: string, update: Partial<User>) {
-    // @TODO (#2090) update org.userIds & permission document if orgId is change for the user
-    await this.db.collection('users').doc(uid).update(update);
-  }
-
-  /**
-   * Fetch all users
-   * @dev If in the future, we need to keep an user list in the state other than members of an org, 
-   * this will be the time to create a userService and to move this method in it.
-   */
-  public async getAllUsers(): Promise<User[]> {
-    const usersSnap = await this.db
-      .collection('users')
-      .get()
-      .toPromise();
-    return usersSnap.docs.map(c => c.data() as User);
-  }
-
-  // TODO THIS IS A QUICK FIX OF MOVIE FINANCING RANK MADE FOR TORONTO, THINK OF A BETTER WAY AFTERWARD
-  //---------------------------
-  //   MOVIE FINANCING RANK
-  //---------------------------
-  public changeRank(rank: string) {
-    this.store.updateProfile({ financing: { rank } });
-  }
 }
