+++ libs/user/src/lib/user/index.ts
@@ -1,4 +1,7 @@
 export * from './account/account.module';
+export * from './+state/user.store';
+export * from './+state/user.service';
+export * from './+state/user.query';
 
 // Component Module
 export * from './components/card/card.module';
